--- satisfaketion-core/src/main/kotlin/io/github/rgbrizzlehizzle/satisfaketion/core/Generator.kt
@@ -1,7 +1,0 @@
-package io.github.rgbrizzlehizzle.satisfaketion.core
-
-import kotlin.random.Random
-
-fun interface Generator<R> {
-  fun generate(seed: Random): R
-}
