--- satisfaketion-core/src/main/kotlin/io/github/rgbrizzlehizzle/satisfaketion/core/Mutator.kt
@@ -1,5 +1,0 @@
-package io.github.rgbrizzlehizzle.satisfaketion.core
-
-fun interface Mutator<R, RR> {
-  fun mutate(generator: Generator<R>): Generator<RR>
-}
