+++ satisfaketion-core/src/main/kotlin/io/github/unredundant/satisfaketion/core/Generator.kt
@@ -1,0 +1,7 @@
+package io.github.unredundant.satisfaketion.core
+
+import kotlin.random.Random
+
+fun interface Generator<R> {
+  fun generate(seed: Random): R
+}
