--- satisfaketion-core/src/test/kotlin/io/github/rgbrizzlehizzle/satisfaketion/core/util/TestModels.kt
@@ -1,4 +1,0 @@
-package io.github.rgbrizzlehizzle.satisfaketion.core.util
-
-data class SimpleDataClass(val a: String, val b: Int)
-data class AnotherSimpleClass(val c: Boolean, val d: String = "hey dude")
