+++ satisfaketion-core/src/test/kotlin/io/github/unredundant/satisfaketion/core/util/TestModels.kt
@@ -1,0 +1,4 @@
+package io.github.unredundant.satisfaketion.core.util
+
+data class SimpleDataClass(val a: String, val b: Int)
+data class AnotherSimpleClass(val c: Boolean, val d: String = "hey dude")
