--- satisfaketion-generators/src/main/kotlin/io/github/rgbrizzlehizzle/satisfaketion/generators/common/Utils.kt
@@ -1,7 +1,0 @@
-package io.github.rgbrizzlehizzle.satisfaketion.generators.common
-
-object Utils {
-  fun getFile(fileName: String): String {
-    return this::class.java.classLoader.getResource(fileName).readText()
-  }
-}
